fix(index): sidenav truncation and separate titles
Only truncate the title text of the article in the side navigation bar
if it's longer then 50 characters. Some articles are shorter then that
so no use adding a misleading ellipse.

A `<br>` element adds a little bit of break between the titles in the
side nav but not enough for quick scanning. Changed the divider to a
`<hr>` tag.