feat(embeddings): save embedded articles 620 at a time
If you're processing thousands or hundreds of thousands of articles it
sucks to lose all the progress if you stop the task. Processing them in
chunks of 20 article batches of 32 at a time means that everything is
getting saved in the database every minute or so. Then you can interrupt
the task at pretty much any time without losing progress.